+++ lib/swagger_frog.dart
@@ -1,3 +1,5 @@
 library;
 
-export 'src/annotations/swagger_frog_annotations.dart';
+export 'src/annotations/openapi_annotations.dart';
+export 'src/middleware/openapi_middleware.dart';
+export 'src/core/api_type_enum.dart';